refactor of random ray filenames
