+++ CMakeLists.txt
@@ -382,11 +382,9 @@ list(APPEND libopenmc_SOURCES
   src/progress_bar.cpp
   src/random_dist.cpp
   src/random_lcg.cpp
-  src/random_ray/iteration.cpp
+  src/random_ray/random_ray_simulation.cpp
   src/random_ray/random_ray.cpp
-  src/random_ray/source_region.cpp
-  src/random_ray/tally_convert.cpp
-  src/random_ray/vtk_plot.cpp
+  src/random_ray/flat_source_domain.cpp
   src/reaction.cpp
   src/reaction_product.cpp
   src/scattdata.cpp
