+++ include/openmc/random_ray/flat_source_domain.h
