+++ include/openmc/random_ray/random_ray_simulation.h
@@ -1,8 +1,7 @@
 #ifndef OPENMC_RANDOM_RAY_ITERATION_H
 #define OPENMC_RANDOM_RAY_ITERATION_H
 
-#include "openmc/random_ray/random_ray.h"
-#include "openmc/random_ray/source_region.h"
+#include "openmc/random_ray/flat_source_domain.h"
 
 namespace openmc {
