--- include/openmc/random_ray/vtk_plot.h
@@ -1,11 +1,0 @@
-#ifndef OPENMC_RANDOM_RAY_VTK_PLOT_H
-#define OPENMC_RANDOM_RAY_VTK_PLOT_H
-
-#include "openmc/random_ray/random_ray.h"
-#include "openmc/random_ray/source_region.h"
-
-namespace openmc {
-
-} // namespace openmc
-
-#endif // OPEMMC_RANMDOM_RAY_VTK_PLOT_H
