+++ src/random_ray/random_ray_simulation.cpp
@@ -1,12 +1,12 @@
-#include "openmc/random_ray/iteration.h"
+#include "openmc/random_ray/random_ray_simulation.h"
+
 #include "openmc/eigenvalue.h"
 #include "openmc/geometry.h"
 #include "openmc/message_passing.h"
 #include "openmc/mgxs_interface.h"
 #include "openmc/output.h"
 #include "openmc/plot.h"
-#include "openmc/random_ray/tally_convert.h"
-#include "openmc/random_ray/vtk_plot.h"
+#include "openmc/random_ray/random_ray.h"
 #include "openmc/simulation.h"
 #include "openmc/source.h"
 #include "openmc/tallies/filter.h"
