--- src/random_ray/vtk_plot.cpp
@@ -1,9 +1,0 @@
-#include "openmc/random_ray/vtk_plot.h"
-#include "openmc/mgxs_interface.h"
-#include "openmc/output.h"
-#include "openmc/plot.h"
-
-namespace openmc {
-
-
-} // namespace openmc
